가격 double -> int 변경

+++ backend/src/main/java/com/caffe/domain/purchase/dto/res/PurchaseDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record PurchaseDto(
         int purchaseId,
         String userEmail,
-        double totalPrice,
+        int totalPrice,
         PurchaseStatus purchaseStatus,
         LocalDateTime purchaseDate
 ) {
